+++ src/communication/azext_communication/aaz/latest/communication/email/domain/sender_username/_create.py
@@ -16,6 +16,9 @@ from azure.cli.core.aaz import *
 )
 class Create(AAZCommand):
     """Create a new SenderUsername resource under the parent Domains resource or update an existing SenderUsername resource.
+
+    :example: Create a sender username
+        az communication email domain sender-username create --domain-name DomainName --email-service-name ResourceName -g ResourceGroup --sender-username SenderUsername --username Username --display-name DisplayName
     """
 
     _aaz_info = {