+++ src/new-relic/azext_new_relic/aaz/latest/new_relic/organization/_list.py
@@ -22,11 +22,13 @@ class List(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2022-07-01",
+        "version": "2024-01-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/providers/newrelic.observability/organizations", "2022-07-01"],
+            ["mgmt-plane", "/subscriptions/{}/providers/newrelic.observability/organizations", "2024-01-01"],
         ]
     }
+
+    AZ_SUPPORT_PAGINATION = True
 
     def _handler(self, command_args):
         super()._handler(command_args)
@@ -45,7 +47,7 @@ class List(AAZCommand):
         _args_schema = cls._args_schema
         _args_schema.location = AAZStrArg(
             options=["--location"],
-            help="Location for NewRelic.",
+            help="Location of NewRelic organization.",
             required=True,
         )
         _args_schema.user_email = AAZStrArg(
@@ -121,7 +123,7 @@ class List(AAZCommand):
                     required=True,
                 ),
                 **self.serialize_query_param(
-                    "api-version", "2022-07-01",
+                    "api-version", "2024-01-01",
                     required=True,
                 ),
             }