+++ src/spring/azext_spring/vendored_sdks/appplatform/v2023_12_01/operations/_api_portal_custom_domains_operations.py
@@ -1,4 +1,4 @@
-# pylint: disable=too-many-lines
+# pylint: disable=too-many-lines,too-many-statements
 # coding=utf-8
 # --------------------------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
@@ -223,7 +223,6 @@ class ApiPortalCustomDomainsOperations:
         :type api_portal_name: str
         :param domain_name: The name of the API portal custom domain. Required.
         :type domain_name: str
-        :keyword callable cls: A custom type or function that will be passed the direct response
         :return: ApiPortalCustomDomainResource or the result of cls(response)
         :rtype: ~azure.mgmt.appplatform.v2023_12_01.models.ApiPortalCustomDomainResource
         :raises ~azure.core.exceptions.HttpResponseError:
@@ -242,23 +241,22 @@ class ApiPortalCustomDomainsOperations:
         api_version: str = kwargs.pop("api_version", _params.pop("api-version", self._api_version or "2023-12-01"))
         cls: ClsType[_models.ApiPortalCustomDomainResource] = kwargs.pop("cls", None)
 
-        request = build_get_request(
+        _request = build_get_request(
             resource_group_name=resource_group_name,
             service_name=service_name,
             api_portal_name=api_portal_name,
             domain_name=domain_name,
             subscription_id=self._config.subscription_id,
             api_version=api_version,
-            template_url=self.get.metadata["url"],
             headers=_headers,
             params=_params,
         )
-        request = _convert_request(request)
-        request.url = self._client.format_url(request.url)
+        _request = _convert_request(_request)
+        _request.url = self._client.format_url(_request.url)
 
         _stream = False
         pipeline_response: PipelineResponse = self._client._pipeline.run(  # pylint: disable=protected-access
-            request, stream=_stream, **kwargs
+            _request, stream=_stream, **kwargs
         )
 
         response = pipeline_response.http_response
@@ -270,13 +268,9 @@ class ApiPortalCustomDomainsOperations:
         deserialized = self._deserialize("ApiPortalCustomDomainResource", pipeline_response)
 
         if cls:
-            return cls(pipeline_response, deserialized, {})
+            return cls(pipeline_response, deserialized, {})  # type: ignore
 
-        return deserialized
-
-    get.metadata = {
-        "url": "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.AppPlatform/Spring/{serviceName}/apiPortals/{apiPortalName}/domains/{domainName}"
-    }
+        return deserialized  # type: ignore
 
     def _create_or_update_initial(
         self,
@@ -284,7 +278,7 @@ class ApiPortalCustomDomainsOperations:
         service_name: str,
         api_portal_name: str,
         domain_name: str,
-        api_portal_custom_domain_resource: Union[_models.ApiPortalCustomDomainResource, IO],
+        api_portal_custom_domain_resource: Union[_models.ApiPortalCustomDomainResource, IO[bytes]],
         **kwargs: Any
     ) -> _models.ApiPortalCustomDomainResource:
         error_map = {
@@ -310,7 +304,7 @@ class ApiPortalCustomDomainsOperations:
         else:
             _json = self._serialize.body(api_portal_custom_domain_resource, "ApiPortalCustomDomainResource")
 
-        request = build_create_or_update_request(
+        _request = build_create_or_update_request(
             resource_group_name=resource_group_name,
             service_name=service_name,
             api_portal_name=api_portal_name,
@@ -320,16 +314,15 @@ class ApiPortalCustomDomainsOperations:
             content_type=content_type,
             json=_json,
             content=_content,
-            template_url=self._create_or_update_initial.metadata["url"],
             headers=_headers,
             params=_params,
         )
-        request = _convert_request(request)
-        request.url = self._client.format_url(request.url)
+        _request = _convert_request(_request)
+        _request.url = self._client.format_url(_request.url)
 
         _stream = False
         pipeline_response: PipelineResponse = self._client._pipeline.run(  # pylint: disable=protected-access
-            request, stream=_stream, **kwargs
+            _request, stream=_stream, **kwargs
         )
 
         response = pipeline_response.http_response
@@ -348,10 +341,6 @@ class ApiPortalCustomDomainsOperations:
             return cls(pipeline_response, deserialized, {})  # type: ignore
 
         return deserialized  # type: ignore
-
-    _create_or_update_initial.metadata = {
-        "url": "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.AppPlatform/Spring/{serviceName}/apiPortals/{apiPortalName}/domains/{domainName}"
-    }
 
     @overload
     def begin_create_or_update(
@@ -383,14 +372,6 @@ class ApiPortalCustomDomainsOperations:
         :keyword content_type: Body Parameter content-type. Content type parameter for JSON body.
          Default value is "application/json".
         :paramtype content_type: str
-        :keyword callable cls: A custom type or function that will be passed the direct response
-        :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :keyword polling: By default, your polling method will be ARMPolling. Pass in False for this
-         operation to not poll, or pass in your own initialized polling object for a personal polling
-         strategy.
-        :paramtype polling: bool or ~azure.core.polling.PollingMethod
-        :keyword int polling_interval: Default waiting time between two polls for LRO operations if no
-         Retry-After header is present.
         :return: An instance of LROPoller that returns either ApiPortalCustomDomainResource or the
          result of cls(response)
         :rtype:
@@ -405,7 +386,7 @@ class ApiPortalCustomDomainsOperations:
         service_name: str,
         api_portal_name: str,
         domain_name: str,
-        api_portal_custom_domain_resource: IO,
+        api_portal_custom_domain_resource: IO[bytes],
         *,
         content_type: str = "application/json",
         **kwargs: Any
@@ -423,18 +404,10 @@ class ApiPortalCustomDomainsOperations:
         :type domain_name: str
         :param api_portal_custom_domain_resource: The API portal custom domain for the create or update
          operation. Required.
-        :type api_portal_custom_domain_resource: IO
+        :type api_portal_custom_domain_resource: IO[bytes]
         :keyword content_type: Body Parameter content-type. Content type parameter for binary body.
          Default value is "application/json".
         :paramtype content_type: str
-        :keyword callable cls: A custom type or function that will be passed the direct response
-        :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :keyword polling: By default, your polling method will be ARMPolling. Pass in False for this
-         operation to not poll, or pass in your own initialized polling object for a personal polling
-         strategy.
-        :paramtype polling: bool or ~azure.core.polling.PollingMethod
-        :keyword int polling_interval: Default waiting time between two polls for LRO operations if no
-         Retry-After header is present.
         :return: An instance of LROPoller that returns either ApiPortalCustomDomainResource or the
          result of cls(response)
         :rtype:
@@ -449,7 +422,7 @@ class ApiPortalCustomDomainsOperations:
         service_name: str,
         api_portal_name: str,
         domain_name: str,
-        api_portal_custom_domain_resource: Union[_models.ApiPortalCustomDomainResource, IO],
+        api_portal_custom_domain_resource: Union[_models.ApiPortalCustomDomainResource, IO[bytes]],
         **kwargs: Any
     ) -> LROPoller[_models.ApiPortalCustomDomainResource]:
         """Create or update the API portal custom domain.
@@ -464,20 +437,9 @@ class ApiPortalCustomDomainsOperations:
         :param domain_name: The name of the API portal custom domain. Required.
         :type domain_name: str
         :param api_portal_custom_domain_resource: The API portal custom domain for the create or update
-         operation. Is either a ApiPortalCustomDomainResource type or a IO type. Required.
+         operation. Is either a ApiPortalCustomDomainResource type or a IO[bytes] type. Required.
         :type api_portal_custom_domain_resource:
-         ~azure.mgmt.appplatform.v2023_12_01.models.ApiPortalCustomDomainResource or IO
-        :keyword content_type: Body Parameter content-type. Known values are: 'application/json'.
-         Default value is None.
-        :paramtype content_type: str
-        :keyword callable cls: A custom type or function that will be passed the direct response
-        :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :keyword polling: By default, your polling method will be ARMPolling. Pass in False for this
-         operation to not poll, or pass in your own initialized polling object for a personal polling
-         strategy.
-        :paramtype polling: bool or ~azure.core.polling.PollingMethod
-        :keyword int polling_interval: Default waiting time between two polls for LRO operations if no
-         Retry-After header is present.
+         ~azure.mgmt.appplatform.v2023_12_01.models.ApiPortalCustomDomainResource or IO[bytes]
         :return: An instance of LROPoller that returns either ApiPortalCustomDomainResource or the
          result of cls(response)
         :rtype:
@@ -512,7 +474,7 @@ class ApiPortalCustomDomainsOperations:
         def get_long_running_output(pipeline_response):
             deserialized = self._deserialize("ApiPortalCustomDomainResource", pipeline_response)
             if cls:
-                return cls(pipeline_response, deserialized, {})
+                return cls(pipeline_response, deserialized, {})  # type: ignore
             return deserialized
 
         if polling is True:
@@ -522,17 +484,15 @@ class ApiPortalCustomDomainsOperations:
         else:
             polling_method = polling
         if cont_token:
-            return LROPoller.from_continuation_token(
+            return LROPoller[_models.ApiPortalCustomDomainResource].from_continuation_token(
                 polling_method=polling_method,
                 continuation_token=cont_token,
                 client=self._client,
                 deserialization_callback=get_long_running_output,
             )
-        return LROPoller(self._client, raw_result, get_long_running_output, polling_method)  # type: ignore
-
-    begin_create_or_update.metadata = {
-        "url": "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.AppPlatform/Spring/{serviceName}/apiPortals/{apiPortalName}/domains/{domainName}"
-    }
+        return LROPoller[_models.ApiPortalCustomDomainResource](
+            self._client, raw_result, get_long_running_output, polling_method  # type: ignore
+        )
 
     def _delete_initial(  # pylint: disable=inconsistent-return-statements
         self, resource_group_name: str, service_name: str, api_portal_name: str, domain_name: str, **kwargs: Any
@@ -551,23 +511,22 @@ class ApiPortalCustomDomainsOperations:
         api_version: str = kwargs.pop("api_version", _params.pop("api-version", self._api_version or "2023-12-01"))
         cls: ClsType[None] = kwargs.pop("cls", None)
 
-        request = build_delete_request(
+        _request = build_delete_request(
             resource_group_name=resource_group_name,
             service_name=service_name,
             api_portal_name=api_portal_name,
             domain_name=domain_name,
             subscription_id=self._config.subscription_id,
             api_version=api_version,
-            template_url=self._delete_initial.metadata["url"],
             headers=_headers,
             params=_params,
         )
-        request = _convert_request(request)
-        request.url = self._client.format_url(request.url)
+        _request = _convert_request(_request)
+        _request.url = self._client.format_url(_request.url)
 
         _stream = False
         pipeline_response: PipelineResponse = self._client._pipeline.run(  # pylint: disable=protected-access
-            request, stream=_stream, **kwargs
+            _request, stream=_stream, **kwargs
         )
 
         response = pipeline_response.http_response
@@ -577,11 +536,7 @@ class ApiPortalCustomDomainsOperations:
             raise HttpResponseError(response=response, error_format=ARMErrorFormat)
 
         if cls:
-            return cls(pipeline_response, None, {})
-
-    _delete_initial.metadata = {
-        "url": "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.AppPlatform/Spring/{serviceName}/apiPortals/{apiPortalName}/domains/{domainName}"
-    }
+            return cls(pipeline_response, None, {})  # type: ignore
 
     @distributed_trace
     def begin_delete(
@@ -598,14 +553,6 @@ class ApiPortalCustomDomainsOperations:
         :type api_portal_name: str
         :param domain_name: The name of the API portal custom domain. Required.
         :type domain_name: str
-        :keyword callable cls: A custom type or function that will be passed the direct response
-        :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :keyword polling: By default, your polling method will be ARMPolling. Pass in False for this
-         operation to not poll, or pass in your own initialized polling object for a personal polling
-         strategy.
-        :paramtype polling: bool or ~azure.core.polling.PollingMethod
-        :keyword int polling_interval: Default waiting time between two polls for LRO operations if no
-         Retry-After header is present.
         :return: An instance of LROPoller that returns either None or the result of cls(response)
         :rtype: ~azure.core.polling.LROPoller[None]
         :raises ~azure.core.exceptions.HttpResponseError:
@@ -634,7 +581,7 @@ class ApiPortalCustomDomainsOperations:
 
         def get_long_running_output(pipeline_response):  # pylint: disable=inconsistent-return-statements
             if cls:
-                return cls(pipeline_response, None, {})
+                return cls(pipeline_response, None, {})  # type: ignore
 
         if polling is True:
             polling_method: PollingMethod = cast(PollingMethod, ARMPolling(lro_delay, **kwargs))
@@ -643,17 +590,13 @@ class ApiPortalCustomDomainsOperations:
         else:
             polling_method = polling
         if cont_token:
-            return LROPoller.from_continuation_token(
+            return LROPoller[None].from_continuation_token(
                 polling_method=polling_method,
                 continuation_token=cont_token,
                 client=self._client,
                 deserialization_callback=get_long_running_output,
             )
-        return LROPoller(self._client, raw_result, get_long_running_output, polling_method)  # type: ignore
-
-    begin_delete.metadata = {
-        "url": "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.AppPlatform/Spring/{serviceName}/apiPortals/{apiPortalName}/domains/{domainName}"
-    }
+        return LROPoller[None](self._client, raw_result, get_long_running_output, polling_method)  # type: ignore
 
     @distributed_trace
     def list(
@@ -668,7 +611,6 @@ class ApiPortalCustomDomainsOperations:
         :type service_name: str
         :param api_portal_name: The name of API portal. Required.
         :type api_portal_name: str
-        :keyword callable cls: A custom type or function that will be passed the direct response
         :return: An iterator like instance of either ApiPortalCustomDomainResource or the result of
          cls(response)
         :rtype:
@@ -692,18 +634,17 @@ class ApiPortalCustomDomainsOperations:
         def prepare_request(next_link=None):
             if not next_link:
 
-                request = build_list_request(
+                _request = build_list_request(
                     resource_group_name=resource_group_name,
                     service_name=service_name,
                     api_portal_name=api_portal_name,
                     subscription_id=self._config.subscription_id,
                     api_version=api_version,
-                    template_url=self.list.metadata["url"],
                     headers=_headers,
                     params=_params,
                 )
-                request = _convert_request(request)
-                request.url = self._client.format_url(request.url)
+                _request = _convert_request(_request)
+                _request.url = self._client.format_url(_request.url)
 
             else:
                 # make call to next link with the client's api-version
@@ -714,14 +655,14 @@ class ApiPortalCustomDomainsOperations:
                         for key, value in urllib.parse.parse_qs(_parsed_next_link.query).items()
                     }
                 )
-                _next_request_params["api-version"] = self._config.api_version
-                request = HttpRequest(
+                _next_request_params["api-version"] = self._api_version
+                _request = HttpRequest(
                     "GET", urllib.parse.urljoin(next_link, _parsed_next_link.path), params=_next_request_params
                 )
-                request = _convert_request(request)
-                request.url = self._client.format_url(request.url)
-                request.method = "GET"
-            return request
+                _request = _convert_request(_request)
+                _request.url = self._client.format_url(_request.url)
+                _request.method = "GET"
+            return _request
 
         def extract_data(pipeline_response):
             deserialized = self._deserialize("ApiPortalCustomDomainResourceCollection", pipeline_response)
@@ -731,11 +672,11 @@ class ApiPortalCustomDomainsOperations:
             return deserialized.next_link or None, iter(list_of_elem)
 
         def get_next(next_link=None):
-            request = prepare_request(next_link)
+            _request = prepare_request(next_link)
 
             _stream = False
             pipeline_response: PipelineResponse = self._client._pipeline.run(  # pylint: disable=protected-access
-                request, stream=_stream, **kwargs
+                _request, stream=_stream, **kwargs
             )
             response = pipeline_response.http_response
 
@@ -746,7 +687,3 @@ class ApiPortalCustomDomainsOperations:
             return pipeline_response
 
         return ItemPaged(get_next, extract_data)
-
-    list.metadata = {
-        "url": "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.AppPlatform/Spring/{serviceName}/apiPortals/{apiPortalName}/domains"
-    }