+++ src/communication/HISTORY.rst
@@ -2,7 +2,11 @@
 
 Release History
 ===============
-1.8.0
+1.9.1
+++++++
+* Add Email communication Domain resource Initiate and Cancel verification operations
+
+1.9.0
 ++++++
 * Adding managed identity support for stable api version 2023-04-01