+++ src/serviceconnector-passwordless/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+1.0.3
+++++++
+* Fix no attribute error
+
 1.0.2
 ++++++
 * Bypass error of `az postgres flexible-server db show`